requirements-dev.txt
-r requirements.txt
jupyter
django-test-without-migrations
debugpy
black
isort
